@@ -4,18 +4,35 @@
   "description": "A simple node themeable webapp creator",
   "main": "index.js",
   "scripts": {
-    "start:dev": "webpack-dev-server --inline --content-base ./public --history-api-fallback",
-    "start:prod": "webpack && node server.js",
+    "start:dev": "babel-node devServer.js",
+    "start:dev2": "webpack-dev-server --inline --content-base ./public --history-api-fallback",
+    "start:prod": "webpack && babel-node server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Oliver Oxenham",
   "license": "ISC",
   "devDependencies": {
     "babel-cli": "^6.5.1",
+    "babel-core": "^6.5.2",
+    "babel-loader": "^6.2.2",
     "babel-preset-es2015": "^6.5.0",
+    "babel-preset-react": "^6.5.0",
     "better-npm-run": "0.0.7",
-    "eslint": "^2.0.0",
+    "css-loader": "^0.23.1",
+    "eslint": "^2.1.0",
+    "eslint-loader": "^1.2.1",
+    "eslint-plugin-react": "^3.16.1",
+    "json-loader": "^0.5.4",
+    "node-sass": "^3.4.2",
+    "react": "^0.14.7",
+    "react-dom": "^0.14.7",
+    "sass-loader": "^3.1.2",
+    "style-loader": "^0.13.0",
     "webpack": "^1.12.13",
-    "webpack-dev-server": "^1.14.1"
+    "webpack-dev-middleware": "^1.5.1",
+    "webpack-hot-middleware": "^2.7.1"
+  },
+  "dependencies": {
+    "express": "^4.13.4"
   }
 }
