@@ -1,23 +1,29 @@
-/*eslint no-console: 1*/
+/*eslint no-console: 0*/
+import Express from 'express';
+import path from 'path';
+import compression from 'compression';
+import config from './config';
 
-var express = require('express');
-var path = require('path');
-var compression = require('compression');
-
-var app = express();
+const app = Express();
 
 // use compression
 app.use(compression());
 
 // serve our static stuff like index.css
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(Express.static(path.join(__dirname)));
 
 // send all requests to index.html so browserHistory in React Router works
-app.get('*', function (req, res) {
+app.get('/', function (req, res) {
   res.sendFile(path.join(__dirname, 'public', 'index.html'));
 });
 
-var PORT = process.env.PORT || 8080;
-app.listen(PORT, function() {
-  console.log('Production Express server running at localhost:' + PORT);
+app.listen(config.port, function() {
+  // console.info('\n' +
+  // '██╗      ██████╗ ██████╗ ███████╗███████╗████████╗ ██████╗ ███╗   ██╗███████╗\n' +
+  // '██║     ██╔═══██╗██╔══██╗██╔════╝██╔════╝╚══██╔══╝██╔═══██╗████╗  ██║██╔════╝\n' +
+  // '██║     ██║   ██║██║  ██║█████╗  ███████╗   ██║   ██║   ██║██╔██╗ ██║█████╗  \n' +
+  // '██║     ██║   ██║██║  ██║██╔══╝  ╚════██║   ██║   ██║   ██║██║╚██╗██║██╔══╝  \n' +
+  // '███████╗╚██████╔╝██████╔╝███████╗███████║   ██║   ╚██████╔╝██║ ╚████║███████╗\n' +
+  // '╚══════╝ ╚═════╝ ╚═════╝ ╚══════╝╚══════╝   ╚═╝    ╚═════╝ ╚═╝  ╚═══╝╚══════╝\n');
+  console.info('----\n==> ✅  %s is running on http://%s:%s', config.app.title, config.host, config.port);
 });
